changes mapping

+++ x-pack/solutions/security/plugins/security_solution/public/common/lib/telemetry/events/rule_deprecation/index.ts
@@ -10,7 +10,7 @@ import { RuleDeprecationEventTypes } from './types';
 
 const countSchema = {
   count: {
-    type: 'long' as const,
+    type: 'integer' as const,
     _meta: {
       description: 'Number of deprecated rules',
       optional: false as const,
